# openapi.json
{
  "components": {
    "schemas": {
      "CreateRestorePointRequest": {
        "properties": {
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "guarantee": {
            "default": true,
            "title": "Guarantee",
            "type": "boolean"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateRestorePointRequest",
        "type": "object"
      },
      "DatabaseTarget": {
        "description": "Exactly one of restore_point / scn / timestamp (spec §3 flashback/database).",
        "properties": {
          "restore_point": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restore Point"
          },
          "scn": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scn"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "title": "DatabaseTarget",
        "type": "object"
      },
      "FinalizeRequest": {
        "properties": {
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          },
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          }
        },
        "title": "FinalizeRequest",
        "type": "object"
      },
      "FlashbackDatabaseRequest": {
        "properties": {
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          },
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "target": {
            "$ref": "#/components/schemas/DatabaseTarget"
          }
        },
        "required": [
          "target"
        ],
        "title": "FlashbackDatabaseRequest",
        "type": "object"
      },
      "FlashbackDropRequest": {
        "properties": {
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "rename_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rename To"
          },
          "table_name": {
            "title": "Table Name",
            "type": "string"
          }
        },
        "required": [
          "owner",
          "table_name"
        ],
        "title": "FlashbackDropRequest",
        "type": "object"
      },
      "FlashbackTableRequest": {
        "properties": {
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "enable_row_movement": {
            "default": false,
            "title": "Enable Row Movement",
            "type": "boolean"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "table_name": {
            "title": "Table Name",
            "type": "string"
          },
          "target": {
            "$ref": "#/components/schemas/FlashbackTarget"
          }
        },
        "required": [
          "owner",
          "table_name",
          "target"
        ],
        "title": "FlashbackTableRequest",
        "type": "object"
      },
      "FlashbackTarget": {
        "description": "Exactly one of scn / timestamp (spec §3 flashback/table).",
        "properties": {
          "scn": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scn"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "title": "FlashbackTarget",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API over DBA-SOP-014 (sop/oracle-flashback-recovery.md)",
    "title": "Oracle Flashback Recovery API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/audit/log": {
      "get": {
        "description": "Audit trail, newest first (SOP §8).",
        "operationId": "audit_log_audit_log_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Log"
      }
    },
    "/flashback/database": {
      "post": {
        "description": "IRREVERSIBLE: rewind the whole database (SOP §5, approval required).\nEnds in READ ONLY validation state; finalize with /flashback/database/finalize.",
        "operationId": "flashback_database_flashback_database_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlashbackDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "把整個資料庫回溯到誤操作之前的狀態，救回大規模誤刪／誤改（不可逆，需審批）"
      }
    },
    "/flashback/database/finalize": {
      "post": {
        "description": "IRREVERSIBLE: OPEN RESETLOGS after manual validation (SOP §5 steps 4-5).",
        "operationId": "finalize_database_flashback_database_finalize_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "人工驗證後正式套用資料庫回溯（OPEN RESETLOGS，不可逆）"
      }
    },
    "/flashback/drop": {
      "post": {
        "description": "Restore a dropped table from the recycle bin (SOP §4.2).",
        "operationId": "flashback_drop_flashback_drop_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlashbackDropRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "從資源回收筒救回被誤刪（DROP）的整張資料表"
      }
    },
    "/flashback/status": {
      "get": {
        "description": "Aggregated precondition checks P1-P4 (SOP §2).",
        "operationId": "flashback_status_flashback_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Flashback Status"
      }
    },
    "/flashback/table": {
      "post": {
        "description": "Rewind a table to a past SCN/timestamp (SOP §4.1). Reversible:\nresponse carries prior_scn for flashing back again.",
        "operationId": "flashback_table_flashback_table_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlashbackTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "把資料表回溯到過去的時間點或 SCN，救回被誤改、誤刪的資料列（可逆）"
      }
    },
    "/health": {
      "get": {
        "description": "Health check.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/recyclebin": {
      "get": {
        "description": "Recycle bin contents, newest drop first (SOP §4.2 step 1).",
        "operationId": "recyclebin_recyclebin_get",
        "parameters": [
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查詢資源回收筒中被刪除（DROP）的資料表，最近刪的排最前"
      }
    },
    "/restore_points": {
      "get": {
        "description": "Existing restore points ordered by SCN (SOP §3.2).",
        "operationId": "list_restore_points_restore_points_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Restore Points"
      },
      "post": {
        "description": "Create a (guaranteed) restore point before a risky change (SOP §3.1).",
        "operationId": "create_restore_point_restore_points_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRestorePointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Restore Point"
      }
    },
    "/restore_points/{name}": {
      "delete": {
        "description": "Drop a restore point after the change is verified (SOP §3.3/§6).",
        "operationId": "drop_restore_point_restore_points__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": true,
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drop Restore Point"
      }
    }
  }
}